+++ dockerfile
@@ -11,7 +11,7 @@ WORKDIR /usr/src/app
 COPY package*.json ./
 RUN npm install
 
-COPY ./src ./src
+COPY ./dist ./src
 
 
 CMD [ "node", "./src/bot.js" ]
