+++ package.json
@@ -5,8 +5,9 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node ./src/bot.js",
-    "devStart": "nodemon ./src/bot.js"
+    "start": "node ./dist/bot.js",
+    "devStart": "nodemon ./dist/bot.js",
+    "build": "tsc"
   },
   "repository": {
     "type": "git",
@@ -27,7 +28,9 @@
     "string-sanitizer": "^2.0.2"
   },
   "devDependencies": {
-    "nodemon": "^2.0.9"
+    "@tsconfig/node17": "^1.0.0",
+    "nodemon": "^2.0.9",
+    "typescript": "^4.7.3"
   },
   "keywords": []
 }
