+++ src/commands/fun/beep.ts
@@ -1,7 +1,9 @@
+import { Message } from "discord.js";
+
 module.exports = {
 	name: 'beep',
 	description: 'Beep!',
-	execute(message) {
+	execute(message: Message) {
 		message.channel.send('Boop');
 	},
 };
