+++ src/commands/fun/ping.ts
@@ -1,3 +1,5 @@
+import { Message } from "discord.js";
+
 module.exports = {
 	name: 'ping',
 	description: 'Ping!',
@@ -6,7 +8,7 @@ module.exports = {
     args: false,
     aliases: ['pinging'],
     usage: '',
-	execute(message) {
+	execute(message: Message) {
 		message.channel.send('Pong');
 	},
 };
