+++ src/data/guild/Config.ts
@@ -1,0 +1,7 @@
+export class Config {
+    prefix = "gg!";
+
+    constructor(prefix = "gg!") {
+        this.prefix = prefix;
+    }
+}
