+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "@tsconfig/node17/tsconfig.json",
+    "compilerOptions": {
+        "outDir": "dist"
+    },
+    "include": ["src"],
+    "exclude": ["node_modules"]
+  }
